#812 CustomerBillingInfo + test
